rename feature_set file
